requirements.txt
beautifulsoup4==4.14.3
blinker==1.9.0
boto2==0.1.0
boto3==1.41.5
Faker==38.2.0
fastapi==0.121.0
huggingface-hub==0.36.0
langchain==1.2.0
langchain-aws==1.1.0
langchain-classic==1.0.0
langchain-cohere==0.5.0
langchain-community==0.4.1
langchain-core==1.2.1
langchain-huggingface==1.2.0
langchain-text-splitters==1.1.0
langchain-weaviate==0.0.6
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.4.59
lxml==6.0.2
Markdown==3.10
markdown-it-py==4.0.0
markdownlit==0.0.7
matplotlib==3.10.7
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.12.0
networkx==3.4.2
openai==2.7.1
pandas==2.3.3
pillow==11.3.0
plotly==6.5.0
pydantic==2.11.10
pydantic-settings==2.12.0
pydantic_core==2.33.2
PyMuPDF==1.26.7
pypdf==6.1.3
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
ruff==0.14.3
safetensors==0.7.0
sentence-transformers==5.2.0
streamlit==1.51.0
tokenizers==0.22.1
torch==2.9.1
transformers==4.57.3
uvicorn==0.38.0
weaviate-client==4.18.3
numpy
cohere